web scraping file for Mercola
Can't seem to get a loop to work to iterate through month values (would be 1 through 11). I can make it work with 1 - it will go to the January page and then the rest of the code works. Just can't seem to get it to go through the rest of the months - I have tried a variety of combinations.